UI  (#17)
* Ajout de nouvelles dépendances Radix UI pour les cases à cocher et les slots, mise à jour de Next.js à la version 15.2.3, et refactorisation des composants de gestion des privilèges et des rôles. Amélioration de l'interface utilisateur avec des mises à jour de styles et de composants, ainsi que des ajustements dans les formulaires et les listes d'utilisateurs.

* Ajout de la dépendance "lucide-react" dans package.json et package-lock.json. Refactorisation des imports de composants pour utiliser la nouvelle structure de répertoires, notamment dans les pages de chargement, de non trouvé, et d'autorisation. Mise à jour des composants d'interface utilisateur pour améliorer la cohérence et l'esthétique de l'application.

* Ajout de paramètres de sensibilité à la casse dans les réglages de cSpell et des fichiers dans .vscode/settings.json pour améliorer la vérification orthographique et la gestion des fichiers.

* Refactorisation des routes API pour la gestion des utilisateurs, mise à jour des schémas Prisma et amélioration de l'authentification. Suppression de l'ancien middleware d'authentification et intégration de nouvelles vérifications de privilèges. Mise à jour des messages d'erreur et des réponses pour une meilleure expérience utilisateur.

* Refactorisation des templates privés et administratifs : suppression des appels aux hooks de gestion des rôles et privilèges dans le modèle privé, ajout de ces appels dans le modèle administratif. Mise à jour des styles de la barre latérale mobile pour améliorer l'interface utilisateur.

* Ajout d'un composant Textarea pour le message dans le gestionnaire de notifications, remplaçant l'ancien Input. Mise à jour de la gestion des événements pour le changement de valeur du message. Amélioration de la logique de fermeture de la barre latérale sur mobile avec un nouveau gestionnaire d'événements.

* Correction du nom de l'application de "Scolitack" à "Scolitrack" dans la configuration de l'application. Ajout d'un nouveau service pour la gestion des notifications backend, incluant des fonctionnalités pour créer, envoyer et gérer les notifications des utilisateurs.

+++ src/components/roles-privileges/RolesList.tsx
@@ -1,7 +1,8 @@
 "use client";
-import { Button } from "@/components/shared/button";
-import { Role, useRoleStore } from "@/context";
-import Card from "../shared/card";
+import { DeleteButton, UpdateButton } from "@/components/ui/button";
+import { useRoleStore } from "@/context/store/RoleStore";
+import { Role } from "@/context/store/types";
+import Card from "../ui/card";
 export const RoleList = () => {
   const { roles } = useRoleStore();
 
@@ -23,7 +24,10 @@ const RoleTable = ({ roles }: { roles: Role[] }) => {
   const { deleteRole, isLoading } = useRoleStore();
 
   const handleDelete = async (id: string) => {
-    await deleteRole(id);
+    const confirm = window.confirm("Voulez-vous vraiment supprimer ce rôle ?");
+    if (confirm) {
+      await deleteRole(id);
+    }
   };
   return (
     <div className="max-w-[1200px] mx-auto">
@@ -62,22 +66,23 @@ const RoleTable = ({ roles }: { roles: Role[] }) => {
                     role.name !== "ADMIN" &&
                     role.name !== "USER" && (
                       <div className="flex gap-2 items-center justify-end">
-                        <Button
-                          variant="accent"
-                          size="sm"
-                          className="bg-transparent border-accent border text-accent hover:text-white"
+                        <UpdateButton
+                          variant="outline"
+                          color="accent"
+                          iconOnly
+                          size="icon-sm"
                         >
                           Modifier
-                        </Button>
-                        <Button
+                        </UpdateButton>
+                        <DeleteButton
+                          variant="outline"
                           onClick={() => handleDelete(role.id)}
                           disabled={isLoading}
-                          variant="destructive"
-                          size="sm"
-                          className="bg-transparent border-red-500 border text-red-500 hover:text-white"
+                          size="icon-sm"
+                          iconOnly
                         >
                           {isLoading ? "en cours..." : "supprimer"}
-                        </Button>
+                        </DeleteButton>
                       </div>
                     )}
                 </td>
